refactor(examples/audiobook-curator): split the operation catalog into feature modules
application.tsx was an 808-line monolith mixing the argv toolkit, shared zod
fragments, sixteen operation declarations, their default executors, and the
bundle tree. Move the catalog into src/operations/ feature modules (evidence,
media-mutation, audible, discovery, output) plus shared cli-arguments.ts and
schemas.ts, leaving application.tsx as composition only. Code is moved
verbatim; operation order, schemas, and behavior are unchanged.
